Add SoundManager and sound systems, improve error handling
* Added a new `SoundManager` class to manage sound effects.
* Modified `ActorManager` to include a `SoundManager` instance.
* Introduced `Comp_SoundEmitter` component for sound emission.
* Updated `SysManager` to include a new `Sys_Sound` system.
* Added `SoundType` enumeration for different sound types.
* Refactored `StateType` enumeration into its own header file.
* Updated various systems to dispatch sound-related messages.
* Improved `ResourceManager` file path handling and added return in `acquireResource`.
* Enhanced `SpriteSheet` class with detailed error messages.
* Added new source and header files to project configuration.
* Enhanced `Sys_InvaderControl` with better random distributions.
* Added friend class declaration in `Sys` for `SysManager` access.
* Updated `Sys_Sound` to handle and play sound-related messages.
* Improved error handling and logging throughout the codebase.

+++ State_Game.cpp
@@ -19,7 +19,7 @@ void State_Game::update(const float& deltaTime)
 {
 	if (m_levelManager.getInvaderCount() <= 0)
 		loadNextLevel();
-	m_fps = 1.0f / deltaTime;
+	m_fps = static_cast<unsigned int>(1.f / deltaTime);
 	// game update
 	m_stateManager->getContext()->m_systemManager->update(deltaTime);
 	// HUD update
